Move panic parsing to zng_task::process::tap
